Added Confirmation for Punish GUI
Added TrustModule - In development

+++ src/com/scorch/core/modules/punish/PunishInventoryListener.java
@@ -14,10 +14,16 @@ import org.bukkit.event.inventory.ClickType;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.event.inventory.InventoryCloseEvent;
 import org.bukkit.event.inventory.InventoryType;
+import org.bukkit.inventory.Inventory;
+import org.bukkit.inventory.ItemStack;
 
 import com.scorch.core.ScorchCore;
 import com.scorch.core.modules.players.CPlayer;
+import com.scorch.core.modules.staff.TrustModule;
+import com.scorch.core.modules.staff.TrustModule.PublicTrust;
+import com.scorch.core.utils.MSG;
 import com.scorch.core.utils.Sounds;
+import com.scorch.core.utils.Utils;
 
 /**
  * Manages Punish GUI
@@ -33,9 +39,13 @@ public class PunishInventoryListener implements Listener {
 	@EventHandler
 	public void onInventoryClick(InventoryClickEvent event) {
 		Player player = (Player) event.getWhoClicked();
+		PunishModule pm = ScorchCore.getInstance().getPunishModule();
+
 		CPlayer cp = ScorchCore.getInstance().getPlayer(player);
 
-		if (event.getCurrentItem() == null || event.getCurrentItem().getType() == Material.AIR)
+		ItemStack item = event.getCurrentItem();
+
+		if (item == null || item.getType() == Material.AIR)
 			return;
 
 		if (!cp.hasTempData("punishing"))
@@ -73,20 +83,39 @@ public class PunishInventoryListener implements Listener {
 				case RIGHT:
 					if (!p.isActive())
 						return;
-					p.remove(player.getName(), cp.getTempString("reason"));
-					player.playSound(player.getLocation(), Sounds.NOTE_BASS_GUITAR.bukkitSound(), 2, 1);
-					player.closeInventory();
+					confirm(player, p, "Confirm punishment &cremoval", "remove");
 					return;
 				case SHIFT_LEFT:
-					ScorchCore.getInstance().getPunishModule().deletePunishment(p);
-					player.playSound(player.getLocation(), Sounds.NOTE_PIANO.bukkitSound(), 2, 1);
-
-					refreshHistory(player, target);
+					confirm(player, p, "Confirm punishment &4deletion", "delete");
 					return;
 				default:
 					return;
 				}
 
+			}
+			return;
+		}
+
+		if ("confirm".equals(cp.getTempData("openInventory"))) {
+			event.setCancelled(true);
+			if (event.getClickedInventory().getType() != InventoryType.CHEST)
+				return;
+			if (item.getType() == Material.GREEN_WOOL) {
+				Punishment p = pm.getPunishment(cp.getTempData("punishremove", UUID.class));
+				String type = cp.getTempString("confirming");
+				String tempReason = cp.getTempString("reason");
+				if (type.equals("remove")) {
+					p.remove(player.getName(), tempReason);
+					player.playSound(player.getLocation(), Sounds.VILLAGER_TRADE.bukkitSound(), 2, 1);
+					player.closeInventory();
+				} else if (type.equals("delete")) {
+					ScorchCore.getInstance().getPunishModule().deletePunishment(p);
+					player.playSound(player.getLocation(), Sounds.VILLAGER_YES.bukkitSound(), 2, 1);
+					player.closeInventory();
+				}
+			} else if (item.getType() == Material.RED_WOOL) {
+				player.playSound(player.getLocation(), Sounds.VILLAGER_NO.bukkitSound(), 2, 1);
+				player.closeInventory();
 			}
 			return;
 		}
@@ -136,15 +165,10 @@ public class PunishInventoryListener implements Listener {
 			case RIGHT:
 				if (!p.isActive())
 					return;
-				player.playSound(player.getLocation(), Sounds.NOTE_BASS_GUITAR.bukkitSound(), 2, 1);
-				p.remove(player.getName(), cp.getTempString("reason"));
-				player.closeInventory();
+				confirm(player, p, "Confirm punishment &cremoval", "remove");
 				return;
 			case SHIFT_LEFT:
-				ScorchCore.getInstance().getPunishModule().deletePunishment(p);
-				player.playSound(player.getLocation(), Sounds.NOTE_PIANO.bukkitSound(), 2, 1);
-
-				refreshPunish(player, target);
+				confirm(player, p, "Confirm punishment &4deletion", "delete");
 				return;
 			default:
 				break;
@@ -156,7 +180,7 @@ public class PunishInventoryListener implements Listener {
 	public void onClose(InventoryCloseEvent event) {
 		Player player = (Player) event.getPlayer();
 		CPlayer cp = ScorchCore.getInstance().getPlayer(player);
-		for (String id : new String[] { "punishing", "reason", "page", "openInventory" })
+		for (String id : new String[] { "punishing", "reason", "page", "openInventory", "trustenum" })
 			cp.removeTempData(id);
 	}
 
@@ -177,8 +201,44 @@ public class PunishInventoryListener implements Listener {
 		CPlayer cp = ScorchCore.getInstance().getPlayer(player);
 		String tempPunish = cp.getTempString("punishing");
 		String tempReason = cp.getTempString("reason");
+
+		TrustModule tm = ScorchCore.getInstance().getModule("TrustModule", TrustModule.class);
+
+		cp.setTempData("trustenum", MSG.color(PublicTrust.get(tm.getTrust(target.getUniqueId())).getColored()));
+
 		player.openInventory(ScorchCore.getInstance().getPunishModule().getPunishGUI(player, target));
 		cp.setTempData("openInventory", "punish");
+		cp.setTempData("punishing", tempPunish);
+		cp.setTempData("reason", tempReason);
+	}
+
+	private void confirm(Player player, Punishment punishment, String message, String removeType) {
+		CPlayer cp = ScorchCore.getInstance().getPlayer(player);
+
+		Inventory inv = Bukkit.createInventory(null, 54, MSG.color(message));
+		inv.setItem(13, punishment.getItem());
+
+		ItemStack confirm = Utils.parseItem(ScorchCore.getInstance().getGui(), "confirm", player),
+				cancel = Utils.parseItem(ScorchCore.getInstance().getGui(), "cancel", player);
+
+		for (int y = 2; y < 5; y++) {
+			for (int x = 1; x < 4; x++) {
+				int slot = 8 + (y * 9) + x;
+				inv.setItem(slot, confirm);
+			}
+			for (int x = 7; x < 10; x++) {
+				int slot = 8 + (y * 9) + x;
+				inv.setItem(slot, cancel);
+			}
+		}
+
+		String tempPunish = cp.getTempString("punishing");
+		String tempReason = cp.getTempString("reason");
+
+		player.openInventory(inv);
+		cp.setTempData("punishremove", punishment.getId());
+		cp.setTempData("openInventory", "confirm");
+		cp.setTempData("confirming", removeType);
 		cp.setTempData("punishing", tempPunish);
 		cp.setTempData("reason", tempReason);
 	}
